move stuff to proper locations
